Remove inject method from factory that extends base factory and update models
