+++ projects/ngx-testing-tools/src/lib/component/test-bed/models/component-callback.model.ts
@@ -1,0 +1,4 @@
+import { EnhancedJasmineCallback } from '../../../common/test-bed/models/enhanced-jasmine-callback.model';
+import { ComponentTools } from './component-tools.model';
+
+export type ComponentCallback<T, I extends {}> = EnhancedJasmineCallback<ComponentTools<T, I>>
