+++ projects/ngx-testing-tools/src/lib/component/test-bed/models/index.ts
@@ -1,4 +1,4 @@
-export { ComponentTestBed } from './component-test-bed.models';
+export { ComponentTestBed } from './component-test-bed.model';
 export { ComponentTestBedOptions } from './component-test-bed-options.model';
 export { ComponentTools } from './component-tools.model';
 export { ComponentExtraOptions } from './component-extra-options.model';
