+++ projects/ngx-testing-tools/src/lib/module/test-bed/models/module-callback.model.ts
@@ -1,0 +1,4 @@
+import { EnhancedJasmineCallback } from '../../../common/test-bed/models/enhanced-jasmine-callback.model';
+import { ModuleTools } from './module-tools.model';
+
+export type ModuleCallback<T, I extends {}> = EnhancedJasmineCallback<ModuleTools<T, I>>
