+++ projects/ngx-testing-tools/src/lib/module/test-bed/models/module-test-bed.model.ts
@@ -1,10 +1,11 @@
-import { EnhancedJasmineCallback } from '../../../common/test-bed/models/enhanced-jasmine-callback.model';
+import { ProviderToken } from '@angular/core';
 import { InjectionStore } from '../../../common/test-bed/store/models/injected-store.model';
+import { NonEmptyString, PrettyMerge } from '../../../shared.model';
 import { ModuleTestBedFactory } from '../module-test-bed-factory';
-import { ModuleTools } from './module-tools.model';
+import { ModuleCallback } from './module-callback.model';
 
 export interface ModuleTestBed<T, I extends InjectionStore = InjectionStore> extends ModuleTestBedFactory<T, I> {
   (assertion: ModuleCallback<T, I['injected']>): jasmine.ImplementationCallback;
-}
 
-export type ModuleCallback<T, I extends {}> = EnhancedJasmineCallback<ModuleTools<T, I>>
+  inject<key extends string, instance>(name: NonEmptyString<key>, token: ProviderToken<instance>): ModuleTestBed<T, InjectionStore<PrettyMerge<I['injected'] & { [k in key]: instance }>>>;
+}
