+++ projects/ngx-testing-tools/src/lib/module/test-bed/module-test-bed-factory.ts
@@ -1,11 +1,10 @@
-import { ProviderToken, Type } from '@angular/core';
+import { Type } from '@angular/core';
 import { assertModuleCtor } from '../../common/assertion/assert-module-ctor';
 import { BaseTestBedFactory } from '../../common/test-bed/base/base-test-bed-factory';
 import { buildJasmineCallback } from '../../common/test-bed/jasmine/jasmine-callback';
 import { InjectionStore } from '../../common/test-bed/store/models/injected-store.model';
-import { NonEmptyString, PrettyMerge } from '../../shared.model';
-import { ModuleTestBed, ModuleTestBedOptions } from './models';
-import { ModuleCallback } from './models/module-test-bed.model';
+import { ModuleTestBedOptions } from './models';
+import { ModuleCallback } from './models/module-callback.model';
 import { buildModuleTools } from './module-tools';
 
 export class ModuleTestBedFactory<ModuleType, Store extends InjectionStore = InjectionStore> extends BaseTestBedFactory<ModuleType, Store> {
@@ -16,10 +15,6 @@ export class ModuleTestBedFactory<ModuleType, Store extends InjectionStore = Inj
   ) {
     assertModuleCtor(rootModule);
     super(rootModule, options);
-  }
-
-  public override inject<key extends string, T>(name: NonEmptyString<key>, token: ProviderToken<T>): ModuleTestBed<ModuleType, InjectionStore<PrettyMerge<Store["injected"] & { [k in key]: T }>>> {
-    return super.inject(name, token) as any;
   }
 
   public override async compile(): Promise<void> {
