+++ projects/ngx-testing-tools/src/lib/pipe/test-bed/models/index.ts
@@ -1,3 +1,3 @@
-export { PipeTestBed } from './pipe-test-bed.models';
+export { PipeTestBed } from './pipe-test-bed.model';
 export { PipeTestBedOptions } from './pipe-test-bed-options.model';
 export { PipeTools } from './pipe-tools.model';
