+++ projects/ngx-testing-tools/src/lib/pipe/test-bed/models/pipe-callback.model.ts
@@ -1,0 +1,4 @@
+import { EnhancedJasmineCallback } from '../../../common/test-bed/models/enhanced-jasmine-callback.model';
+import { PipeTools } from './pipe-tools.model';
+
+export type PipeCallback<T, I extends {}> = EnhancedJasmineCallback<PipeTools<T, I>>
