+++ projects/ngx-testing-tools/src/lib/service/test-bed/models/service-callback.model.ts
@@ -1,0 +1,4 @@
+import { EnhancedJasmineCallback } from '../../../common/test-bed/models/enhanced-jasmine-callback.model';
+import { ServiceTools } from './service-tools.model';
+
+export type ServiceCallback<T, I extends {}> = EnhancedJasmineCallback<ServiceTools<T, I>>
